backend/requirements.txt
django
python-decouple
dj-database-url
mysqlclient
djangorestframework
django-cors-headers
djangorestframework-simplejwt
django-filter
reportlab
hypothesis
razorpay
django-ratelimit
pillow
bleach
python-magic
django-redis
redis
